refactor: take dto layer to upper layer from api layer
