+++ src/main/kotlin/com/wafflytime/user/auth/dto/AuthToken.kt
@@ -1,3 +1,3 @@
-package com.wafflytime.user.auth.api.dto
+package com.wafflytime.user.auth.dto
 
 data class AuthToken(val accessToken: String, val refreshToken: String)
