+++ src/main/kotlin/com/wafflytime/user/auth/dto/LoginRequest.kt
@@ -1,12 +1,10 @@
-package com.wafflytime.user.auth.api.dto
+package com.wafflytime.user.auth.dto
 
 import jakarta.validation.constraints.NotBlank
 
-data class SignUpRequest(
+data class LoginRequest(
     @field:NotBlank
     val id: String,
     @field:NotBlank
     val password: String,
-    @field:NotBlank
-    val nickname: String,
 )
