+++ src/main/kotlin/com/wafflytime/user/auth/dto/SignUpRequest.kt
@@ -1,10 +1,12 @@
-package com.wafflytime.user.auth.api.dto
+package com.wafflytime.user.auth.dto
 
 import jakarta.validation.constraints.NotBlank
 
-data class LoginRequest(
+data class SignUpRequest(
     @field:NotBlank
     val id: String,
     @field:NotBlank
     val password: String,
+    @field:NotBlank
+    val nickname: String,
 )
