+++ src/main/kotlin/com/wafflytime/user/auth/dto/SocialSignUpRequest.kt
@@ -1,4 +1,4 @@
-package com.wafflytime.user.auth.api.dto
+package com.wafflytime.user.auth.dto
 
 import jakarta.validation.constraints.NotBlank
