+++ src/main/kotlin/com/wafflytime/user/info/service/UserService.kt
@@ -5,16 +5,16 @@ import com.wafflytime.post.database.PostRepository
 import com.wafflytime.post.database.ScrapRepository
 import com.wafflytime.post.dto.PostResponse
 import com.wafflytime.user.info.api.dto.DeleteScrapResponse
-import com.wafflytime.user.info.api.dto.UpdateUserInfoRequest
 import com.wafflytime.user.info.api.dto.UploadProfileImageRequest
-import com.wafflytime.user.info.api.dto.UserInfo
 import com.wafflytime.user.info.database.UserEntity
 import com.wafflytime.user.info.database.UserRepository
+import com.wafflytime.user.mail.dto.VerifyEmailRequest
+import com.wafflytime.user.info.dto.UpdateUserInfoRequest
+import com.wafflytime.user.info.dto.UserInfo
 import com.wafflytime.user.info.exception.LoginIdConflict
 import com.wafflytime.user.info.exception.MailConflict
 import com.wafflytime.user.info.exception.NicknameConflict
 import com.wafflytime.user.info.exception.UserNotFound
-import com.wafflytime.user.mail.api.dto.VerifyEmailRequest
 import jakarta.transaction.Transactional
 import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.domain.Page
