+++ src/main/kotlin/com/wafflytime/user/mail/dto/VerifyEmailCode.kt
@@ -1,3 +1,3 @@
-package com.wafflytime.user.mail.api.dto
+package com.wafflytime.user.mail.dto
 
 data class VerifyEmailCode(val code: String)
